backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
transformers==4.44.2
torch==2.4.1
pytest==8.3.3
httpx==0.27.2
